requirements.txt
python-binance  
pandas  
ta  
requests  
cryptography 
